+++ apps/evm/src/components/Chip/index.tsx
@@ -1,4 +1,3 @@
-/** @jsxImportSource @emotion/react */
 import { Typography } from '@mui/material';
 import { useMemo } from 'react';