+++ apps/evm/src/components/Table/index.tsx
@@ -1,4 +1,3 @@
-/** @jsxImportSource @emotion/react */
 import MuiTable from '@mui/material/Table';
 import MuiTableBody from '@mui/material/TableBody';
 import MuiTableCell from '@mui/material/TableCell';