+++ apps/evm/src/containers/Layout/store.ts
@@ -1,7 +1,5 @@
 import { create } from 'zustand';
 
-import { createStoreSelectors } from 'utilities';
-
 interface State {
   isScrollToTopVisible: boolean;
   setScrollToTopVisible: (isScrollToTopVisible: State['isScrollToTopVisible']) => void;
@@ -9,11 +7,9 @@ interface State {
   setIsCloseToBottom: (isCloseToBottom: State['isCloseToBottom']) => void;
 }
 
-const useStore = create<State>()(set => ({
+export const useStore = create<State>()(set => ({
   isScrollToTopVisible: false,
   setScrollToTopVisible: (v: boolean) => set({ isScrollToTopVisible: v }),
   isCloseToBottom: false,
   setIsCloseToBottom: (v: boolean) => set({ isCloseToBottom: v }),
 }));
-
-export const store = createStoreSelectors(useStore);