+++ apps/evm/src/pages/Governance/ProposalList/CreateProposalModal/UploadOrManualProposal/index.tsx
@@ -1,4 +1,3 @@
-/** @jsxImportSource @emotion/react */
 import { Box } from '@mui/material';
 import { useRef } from 'react';